Package-internal documentation

Package-internal documentation

This page lists all the documented internals of the SeaIce module.

Index

A list of all internal documentation sorted by module.

SeaIce.bilinearInterpolation!

SeaIce.copyGridSortingInfo!

SeaIce.curl

SeaIce.findOverlap

SeaIce.writeGridVTK

Internal Interface

SeaIce.bilinearInterpolation! — Method.

bilinearInterpolation(field, x_tilde, y_tilde, i, j, k, it)

Use bilinear interpolation to interpolate a staggered grid to an arbitrary position in a cell. Assumes south-west convention, i.e. (i,j) is located at the south-west (-x, -y)-facing corner.

Arguments

field::Array{Float64, 4}: a scalar field to interpolate from

x_tilde::Float64: x point position [0;1]

y_tilde::Float64: y point position [0;1]

i::Int: i-index of cell containing point

j::Int: j-index of scalar field to interpolate from

it::Int: time step from scalar field to interpolate from

source

SeaIce.copyGridSortingInfo! — Method.

Copy ice floe related information from ocean to atmosphere grid. This is useful when the two grids are of identical geometry, meaning only only one sorting phase is necessary.

source

SeaIce.curl — Function.

curl(grid, x_tilde, y_tilde, i, j, k, it)

Use bilinear interpolation to interpolate curl value for a staggered velocity grid to an arbitrary position in a cell. Assumes south-west convention, i.e. (i,j) is located at the south-west (-x, -y)-facing corner.

Arguments

grid::Any: grid for which to determine curl

x_tilde::Float64: x point position [0;1]

y_tilde::Float64: y point position [0;1]

i::Int: i-index of cell containing point

j::Int: j-index of scalar field to interpolate from

it::Int: time step from scalar field to interpolate from

source

SeaIce.findOverlap — Method.

position_ij is the inter-grain position vector, and can be found with interIceFloePositionVector().

source

SeaIce.writeGridVTK — Method.

Write a VTK file to disk containing all ocean data in the simulation in a structured grid (file type .vts). These files can be read by ParaView and can be visualized by applying a Glyph filter. This function is called by writeVTK().

source